backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Redis 缓存
redis==5.0.1
hiredis==2.2.3

# 认证授权
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP 客户端
httpx==0.25.2
aiofiles==23.2.1

# 异步任务队列
celery==5.3.4
flower==2.0.1

# 图像处理
Pillow==10.1.0
opencv-python==4.8.1.78
pydicom==2.4.3

# AI/ML 相关
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# 工具库
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
typer==0.9.0

# 文件处理
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7

# 邮件发送
fastapi-mail==1.4.1
jinja2==3.1.2

# CORS 和安全
fastapi-cors==0.0.6
slowapi==0.1.9

# 配置管理
python-dotenv==1.0.0
dynaconf==3.2.4

# 数据库连接池
asyncpg==0.29.0
aiomysql==0.2.0

# 缓存装饰器
cachetools==5.3.2
diskcache==5.6.3

# 时间处理
arrow==1.3.0
pendulum==2.1.2

# 字符串处理
unidecode==1.3.7
slugify==0.0.1

# 网络请求
requests==2.31.0
urllib3==2.1.0

# JSON 处理
orjson==3.9.10
ujson==5.8.0

# 压缩和解压
# zipfile和tarfile是Python内置模块，无需安装

# 系统监控
psutil==5.9.6
py-cpuinfo==9.0.0

# 地理位置
geopy==2.4.1
geoip2==4.7.0

# 二维码生成
qrcode[pil]==7.4.2
python-barcode==0.15.1

# PDF 处理
PyPDF2==3.0.1
pdfplumber==0.10.3

# Excel 处理
pandas==2.1.4
xlsxwriter==3.1.9

# 加密解密
cryptography==41.0.7
pycryptodome==3.19.0

# 正则表达式
regex==2023.10.3

# 环境变量
environs==10.0.0

# 版本管理
packaging==23.2
semantic-version==2.10.0
